+++ src/main/java/kr/where/backend/group/dto/groupmember/ResponseGroupMemberDTO.java
@@ -13,7 +13,7 @@ public class ResponseGroupMemberDTO {
 
     private Long groupId;
     private String groupName;
-    private Long memberId;
+    private Integer memberId;
     private String comment;
     private String memberIntraName;
     private String customLocation;
@@ -24,7 +24,7 @@ public class ResponseGroupMemberDTO {
 
 
     @Builder
-    public ResponseGroupMemberDTO(@NotNull Long groupId, String groupName, @NotNull Long memberId, String comment, String memberIntraName, String customLocation, String imacLocation, String clusterLocation, boolean inCluster, String image) {
+    public ResponseGroupMemberDTO(@NotNull Long groupId, String groupName, @NotNull Integer memberId, String comment, String memberIntraName, String customLocation, String imacLocation, String clusterLocation, boolean inCluster, String image) {
         this.groupId = groupId;
         this.groupName = groupName;
         this.memberId = memberId;
